Move database connector classes to new database.connector package.
