+++ src/main/java/cloudgene/mapred/database/connector/MySqlConnector.java
@@ -15,7 +15,7 @@
  * along with this program; if not, see <http://www.gnu.org/licenses/>.
  ******************************************************************************/
 
-package cloudgene.mapred.database.util.mysql;
+package cloudgene.mapred.database.connector;
 
 import java.io.BufferedReader;
 import java.io.DataInputStream;
@@ -33,8 +33,6 @@ import org.apache.commons.dbcp.BasicDataSource;
 import org.apache.commons.dbutils.DbUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import cloudgene.mapred.database.util.AbstractDatabaseConnector;
 
 public class MySqlConnector extends AbstractDatabaseConnector {
