+++ src/main/java/cloudgene/mapred/server/Application.java
@@ -17,8 +17,8 @@ import org.slf4j.LoggerFactory;
 import cloudgene.mapred.database.dao.TemplateDao;
 import cloudgene.mapred.database.updates.BcryptHashUpdate;
 import cloudgene.mapred.database.util.Database;
-import cloudgene.mapred.database.util.DatabaseConnector;
-import cloudgene.mapred.database.util.DatabaseConnectorFactory;
+import cloudgene.mapred.database.connector.DatabaseConnector;
+import cloudgene.mapred.database.connector.DatabaseConnectorFactory;
 import cloudgene.mapred.database.util.DatabaseUpdater;
 import cloudgene.mapred.database.util.Fixtures;
 import cloudgene.mapred.jobs.PersistentWorkflowEngine;
